requirements/requirements-web.txt
# Web-specific dependencies
-r requirements-base.txt

# REST framework
djangorestframework>=3.14,<4.0
drf-spectacular>=0.26.0,<1.0

# Storage
django-storages>=1.13,<2.0
minio>=7.1,<8.0

# AI and vector search
crewai>=0.28,<1.0
llama-index>=0.8,<1.0
pgvector>=0.2,<1.0
langfuse>=2.0,<3.0

# Development tools
django-debug-toolbar>=4.1,<5.0
django-extensions>=3.2,<4.0

# Monitoring (if enabled via environment variable)
-r requirements-monitoring.txt
